refactor: replace proxy requests with direct calls to GeminiInteractionClient in deep research API
