--- chat_shell/chat_shell/clients/__init__.py
@@ -1,9 +1,0 @@
-# SPDX-FileCopyrightText: 2025 Weibo, Inc.
-#
-# SPDX-License-Identifier: Apache-2.0
-
-"""HTTP clients for external services."""
-
-from shared.clients.gemini_interaction import GeminiInteractionClient
-
-__all__ = ["GeminiInteractionClient"]
